test: Reservation 테스트 분리 및 예약 상태 값 검증 테스트 추가
